Data import function.

It reads a file with data exported from AAA. The data are automatically transformed from a wide to a long format (each row has values of X or Y axes for each fan line). The imported tibble can then be used for plotting and statistical analysis.

read_aaa_data(file, column_names, fan_lines, coordinates, na_rm, format)

Arguments

file
The file or files with AAA data.

column_names
The names of the columns without including the splines columns.

fan_lines
The number of fan lines.

coordinates
A string specifying the coordinate system. Possible values are "cartesian" and "polar".

na_rm
Remove NAs.

format
A string specifying the data format. Possible values are "long" and "wide" (the default is "long").

Value

An object of class tbl_df-class (a tibble).